+++ matmul/include/matmul/matmul_extend_adown.cuh
@@ -52,8 +52,8 @@ namespace hsys {
 template <MatrixK MatrixT = Matrix<float>>
 void matmul_extend_adown(MatrixT& c, const MatrixT& a, const MatrixT& b) {
   constexpr auto block = dim3(16, 16);
-  const auto grid
-      = dim3(std::ceil(c.ncols() / block.x), std::ceil(c.nrows() / (block.y * 2)));
+  const auto grid = dim3(std::ceil(c.ncols() / block.x),
+      std::ceil(static_cast<float>(c.nrows()) / (block.y * 2)));
   hsys::kernels::matmul_extend_adown<16><<<grid, block>>>(c.view(), a.view(), b.view());
 }
